Separating components. Moved header and footer to the main component - App.svelte
