+++ src/global-vars.scss
@@ -45,7 +45,8 @@ $hover-card-box-shadow: #0000001f;
 $button-hover: #f1f3f4;
 
 $content-max-width: 800px;
-$header-height: 150px;
+$header-height: 80px;
+$header-search-bar-height: 70px;
 $footer-height: 120px;
 $minimum-body-margin: 20px;
 $media-breakpoint: 600px;
